Refactor key names in app and in core
- Key names are just part of the key generation messages
- Key generation is multiple stages in the UI

+++ frostsnapp/native/src/api.rs
@@ -751,7 +751,7 @@ impl Coordinator {
 
     pub fn generate_new_key(
         &self,
-        threshold: usize,
+        threshold: u16,
         devices: Vec<DeviceId>,
         key_name: String,
         event_stream: StreamSink<KeyGenState>,
